requirements.txt
-e ./media_crawler
openai>=1.12.0
pillow>=10.0.0
numpy>=1.26.0
python-dotenv>=0.19.0
playwright>=1.42.0